[Back] Etape 4 Messages
	modified:   back/app/Http/Controllers/MessageController.php
	modified:   back/app/Http/Requests/MessageRequest.php
	modified:   back/app/Http/Resources/MessageResource.php
	modified:   back/app/Models/Message.php
	modified:   back/app/Policies/MessagePolicy.php
	modified:   back/database/factories/MessageFactory.php
	new file:   back/database/migrations/2020_05_17_024449_add_link_and_soft_delete_to_messages_table.php
	modified:   back/database/seeds/MessagesTableSeeder.php

	modified:   back/app/Http/Controllers/MessageController.php
	modified:   back/app/Http/Requests/MessageRequest.php
	modified:   back/app/Http/Resources/MessageResource.php
	modified:   back/app/Models/Message.php
	modified:   back/app/Policies/MessagePolicy.php
	modified:   back/database/factories/MessageFactory.php
	new file:   back/database/migrations/2020_05_17_024449_add_link_and_soft_delete_to_messages_table.php
	modified:   back/database/seeds/MessagesTableSeeder.php

+++ back/app/Http/Controllers/MessageController.php
@@ -12,13 +12,15 @@ class MessageController extends Controller
 {
     public function index()
     {
+        $this->authorize('index', Message::class);
         return response()->json(MessageResource::collection(Message::all()));
     }
 
     public function store(MessageRequest $request)
     {
+        $this->authorize('store', Message::class);
         $message = Message::create(
-            $request->only(['title','content'])
+            $request->only(['title', 'content', 'link'])
         );
 
         return response()->json(MessageResource::make($message), JsonResponse::HTTP_CREATED);
@@ -27,7 +29,8 @@ class MessageController extends Controller
     public function update(MessageRequest $request,$message_id)
     {
         $message = Message::findOrFail($message_id);
-        $message->update($request->only(['title','content']));
+        $this->authorize('update', $message);
+        $message->update($request->only(['title', 'content', 'link']));
 
         return response()->json(MessageResource::make($message), JsonResponse::HTTP_CREATED);
     }
@@ -36,6 +39,7 @@ class MessageController extends Controller
     public function destroy($message_id)
     {
         $message = Message::findOrFail($message_id);
+        $this->authorize('destroy', $message);
         $message->delete();
 
         return response()->json(null, JsonResponse::HTTP_NO_CONTENT);
